+++ pigeons/messages.dart
@@ -2,16 +2,15 @@ import 'package:pigeon/pigeon.dart';
 
 @ConfigurePigeon(
   PigeonOptions(
-    kotlinOptions: KotlinOptions(
-      package: 'fr.g123k.install_referrer'
-    ),
-    swiftOptions: SwiftOptions(),
-    dartOptions: DartOptions(
-    ),
-    dartOut: './lib/src/private/pigeon_api.dart',
-    swiftOut: './ios/Classes/SwiftInstallReferrerPigeon.swift',
-    kotlinOut: './android/src/main/kotlin/fr/g123k/install_referrer/InstallReferrerPigeon.kt'
-  ),
+      kotlinOptions: KotlinOptions(
+        package: 'com.allwinsolutions.flutter_install_referrer'
+      ),
+      swiftOptions: SwiftOptions(),
+      dartOptions: DartOptions(),
+      dartOut: './lib/src/private/pigeon_api.dart',
+      swiftOut: 'ios/flutter_install_referrer/Sources/flutter_install_referrer/SwiftInstallReferrerPigeon.swift',
+      kotlinOut:
+          './android/src/main/kotlin/com/allwinsolutions/flutter_install_referrer/InstallReferrerPigeon.kt'),
 )
 class IRInstallationReferrer {
   final IRInstallationType? type;